third_party/blink/web_tests/external/wpt/css/css-values/calc-infinity-nan-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property width value 'calc(infinity * 1px - infinity * 1%)'
  assert_equals: expected "0px" but got "3.35544e+07px"
[FAIL] Property width value 'calc(infinity * 1px - max(infinity * 1%, 0%))'
  assert_equals: expected "0px" but got "3.35544e+07px"
Harness: the test ran to completion.
